+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   for KiCad 6.
 - Schematic:
   - Support for SHEETNAME, SHEETFILE and SHEETPATH in the title block
+- Globals:
+  - `schematic_sheet_name_workaround` to workaround SHEETNAME on page 1
+    (See #933)
 - Variants:
   - pre_transform filter to KiCad variants
 - E/DRC:
@@ -35,6 +38,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `collect_images`: To just avoid collecting images, enabled by
     default when generating a CSV. Avoids unintended infinite loops.
     (See #931)
+- *SCH Print
+  - `title_propagate`: The title is not only set for page 1 but also for
+    all pages.
 
 ## Fixed
 - Warnings: